+++ tests/unit/middleware/requestIdentity.js
@@ -2,6 +2,7 @@ var sinon     = require('sinon');
 var chai      = require('chai');
 var sinonChai = require("sinon-chai");
 
+var Service         = require('../../../lib/service.js');
 var requestIdentity = require('../../../lib/middleware/requestIdentity.js');
 var AppManager      = require('../../../lib/express/appManager.js');
 var Config          = require('../mocks/config.js');
@@ -14,10 +15,10 @@ chai.should();
 describe('requestIdentity middleware', function() {
 
     before(function() {
-        this.models = {};
         this.config = new Config();
 
-        this.appManager = new AppManager(this.models);
+        this.service = new Service(this.config);
+        this.appManager = this.service.appManager;
         var app = this.app = this.appManager.buildApp(this.config, {name: '1'});
 
         this.res = {};